--- internal/types/to_string.go
@@ -1,34 +1,0 @@
-// Copyright 2021 FerretDB Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package types
-
-import "fmt"
-
-// String is needed to conveniently inform about type or value
-// mismatch when returning an error. For example,
-// when informing that `int32` is an invalid type, in the
-// error message we would like to get `int` and not `int32`.
-func String(v any) string {
-	switch v.(type) {
-	case NullType:
-		return "null"
-	case int32:
-		return "int"
-	case int64, float64:
-		return "double"
-	default:
-		return fmt.Sprintf("%s", v)
-	}
-}
